Removed generic event message and digitiser_event_message
